requirements.txt
Flask
requests==2.26.0
python-dotenv==1.0.0